Add helper functions for database cleanup
